moved automata shoehorning to separate file
