+++ mathbot/tests/conftest.py
@@ -1,6 +1,3 @@
-import pytest
-
-
 def pytest_addoption(parser):
     parser.addoption(
         "--run-automata",
